Validação no Controller

+++ app/Http/Controllers/ProjectController.php
@@ -33,7 +33,7 @@ class ProjectController extends Controller
 
     public function index()
     {
-        return $this->repository->all();
+        return $this->repository->findWhere(['owner_id' => \Authorizer::getResourceOwnerId()]);
     }
 
     public function store(Request $request)
@@ -43,18 +43,33 @@ class ProjectController extends Controller
 
     public function show($id)
     {
+        if($this->checkProjectOwner($id)==false){
+            return ['Erro' => 'Acesso Negado'];
+        }
         return $this->repository->find($id);
     }
 
 
     public function update(Request $request, $id)
     {
+        if($this->checkProjectOwner($id)==false){
+            return ['Erro' => 'Acesso Negado'];
+        }
         return $this->service->update($request->all(),$id);
 
     }
 
     public function destroy($id)
     {
+        if($this->checkProjectOwner($id)==false){
+            return ['Erro' => 'Acesso Negado'];
+        }
         return $this->repository->delete($id);
+    }
+
+    private function checkProjectOwner($projectId){
+        $userId = \Authorizer::getResourceOwnerId();
+
+        if ($this->repository->isOwner($projectId, $userId));
     }
 }
